Rename MonoSymbol to Symbol, widen module_idx u16→u32, deduplicate Symbol definition
- Rename MonoSymbol to Symbol across all 20 files
- Widen module_idx from u16 to u32, packed struct from u48 to u64
- Single Symbol definition in MIR.zig, re-exported by MonoIR.zig via new mono→mir dependency
- Add comptime size/alignment assertions (8 bytes, 8-byte aligned)
- Fix mir/Lower.zig findModuleForOrigin: use self.current_module_idx directly
- Fix mono/Lower.zig findModuleForOrigin: callers pass source_module_idx, eliminating pointer scan
- Replace magic sentinel diagnostics with typed error variants:
  runtime_err_can (from CIR), runtime_err_type, runtime_err_ellipsis, runtime_err_anno_only
- Change e_lookup_pending/e_lookup_required to unreachable (compiler bug if reached)

+++ src/eval/dev_evaluator.zig
@@ -8,7 +8,7 @@
 //! 5. Generating native machine code (x86_64/aarch64)
 //! 6. Executing the generated code
 //!
-//! Code generation uses Mono IR with globally unique MonoSymbol references,
+//! Code generation uses Mono IR with globally unique Symbol references,
 //! eliminating cross-module index collisions.
 
 const std = @import("std");
@@ -666,7 +666,7 @@ pub const DevEvaluator = struct {
         defer mono_store.deinit();
 
         // Find the module index for this module
-        var module_idx: u16 = 0;
+        var module_idx: u32 = 0;
         for (all_module_envs, 0..) |env, i| {
             if (env == module_env) {
                 module_idx = @intCast(i);
