+++ src/snapshot_tool/main.zig
@@ -3891,7 +3891,7 @@ fn processDevObjectSnapshot(
     defer layout_store.deinit();
 
     // 8. Find app module index and lower to Mono IR
-    var app_module_idx: ?u16 = null;
+    var app_module_idx: ?u32 = null;
     for (modules, 0..) |mod, i| {
         if (mod.is_app) {
             app_module_idx = @intCast(i + 1);
